# store: bucket: reuse buffers for Series() responses