--- a/pkg/store/bucket_test.go
+++ b/pkg/store/bucket_test.go
@@ -1482,6 +1482,11 @@ func TestBucketSeries_OneBlock_InMemIndexCacheSegfault(t *testing.T) {
 		queryGate:            gate.NewNoop(),
 		chunksLimiterFactory: NewChunksLimiterFactory(0),
 		seriesLimiterFactory: NewSeriesLimiterFactory(0),
+		respBufferPool: &sync.Pool{
+			New: func() interface{} {
+				return &bytes.Buffer{}
+			},
+		},
 	}
 
 	t.Run("invoke series for one block. Fill the cache on the way.", func(t *testing.T) {