@GiedriusS 2021-08-09
v2 here @bwplotka @yeya24 https://github.com/thanos-io/thanos/pull/4535